CID 720453: Don't cast, so we restrict the range of values to the minimum and maximum indecies of the array
Hopefully coverity is smart enough to figure that out... Otherwise I guess we're checking the that uint8_t can't be bigger than UINT8_MAX :/

+++ src/lib/util/pair.c
@@ -2532,7 +2532,7 @@ FR_TOKEN fr_pair_raw_from_str(char const **ptr, VALUE_PAIR_RAW *raw)
 		 *	out which operators come after the attribute
 		 *	name.  Yes, our "lexer" is bad.
 		 */
-		if (!fr_dict_attr_allowed_chars[(unsigned int) *t]) {
+		if (!fr_dict_attr_allowed_chars[*t]) {
 			break;
 		}
